updating the dc level GRN - Disable functionality

+++ ANGULAR_WMS/app/scripts/controllers/inbound/raise_pr/raise_purchase_request.js
@@ -857,7 +857,7 @@ function ServerSideProcessingCtrl($scope, $http, $q, $state, $rootScope, $compil
       var keepGoing = true
       angular.forEach(elem, function(list_obj) {
         if (list_obj['name'] == 'order_quantity') {
-          if (parseInt(list_obj['value']) <= 0) {
+          if (parseFloat(list_obj['value']) <= 0) {
             keepGoing = false
           }
         }
@@ -1378,7 +1378,7 @@ function ServerSideProcessingCtrl($scope, $http, $q, $state, $rootScope, $compil
       var keepGoing = true
       angular.forEach(elem, function(list_obj) {
         if (list_obj['name'] == 'order_quantity') {
-          if (parseInt(list_obj['value']) <= 0) {
+          if (parseFloat(list_obj['value']) <= 0) {
             keepGoing = false
           }
         }
